Moves all non index view methods from home to new Gen controller, implemetns KIC controller inheritance
