+++ cure/Views/Shared/_Layout.cshtml
@@ -26,7 +26,7 @@
 
     <footer class="border-top footer text-muted">
         <div class="container">
-            &copy; 2024 - KIC Events LLC - <a href="https://www.kicevents.com/Home/Privacy">Privacy</a> - <a href="https://www.kicevents.com/Home/Contact">Contact</a>
+            &copy; 2024 - KIC Events LLC - <a href="https://www.kicevents.com/Gen/Privacy">Privacy</a> - <a href="https://www.kicevents.com/Gen/Contact">Contact</a>
         </div>
     </footer>
     <script src="~/lib/jquery/dist/jquery.min.js"></script>
