+++ kicweb/Views/Gen/Club425.cshtml
@@ -1,5 +1,5 @@
 @page
-@model KiCWeb.Views.Home.Club425
+
 
 <div class="view-body">
     <div class="row text-center">
